Created getters and a module to store the flow
