--- bin/str_content.rb
@@ -1,18 +1,0 @@
-def greetings
-  Gem.win_platform? ? (system 'cls') : (system 'clear')
-  puts
-  puts 'Welcome to Kasuhira Price Scrapping system, how may I be of assistance ?'
-  puts
-  puts '     _________'
-  puts '    / ======= \\'
-  puts '   / __________\\'
-  puts '  | ___________ |'
-  puts '  | | -       | |'
-  puts '  | |         | |'
-  puts '  | |_________| |_____________________________'
-  puts '  \=____________/   Developed by Kasuhira     )'
-  puts '  / """"""""""" \                            /'
-  puts " / ::::::::::::: \\                      =D -'"
-  puts '|_________________|'
-puts
-end
